feat: update

+++ app/src/main/java/ru/vsu/zmaev/a4rotor/data/model/point/PointData.kt
@@ -1,6 +1,6 @@
 package ru.vsu.zmaev.a4rotor.data.model.point
 
 data class PointData (
-    val locationX: Float,
-    val locationY: Float
+    val locationX: Double,
+    val locationY: Double
 )
